refactor: combine vault actions functions into single file
