+++ utils/actions.ts
@@ -1,22 +1,85 @@
 import assert from 'assert';
-import {erc20ABI} from 'wagmi';
+import {erc20ABI, readContract} from '@wagmi/core';
 import VAULT_ABI from '@yearn-finance/web-lib/utils/abi/vault.abi';
 import {toAddress} from '@yearn-finance/web-lib/utils/address';
+import {MAX_UINT_256} from '@yearn-finance/web-lib/utils/constants';
 
-import {assertAddress, handleTx} from './toWagmiProvider';
+import {assertAddress, handleTx, toWagmiProvider} from './toWagmiProvider';
 import FACTORY_KEEPER_ABI from './ABI/factoryKeeper.abi';
 
+import type {Connector} from 'wagmi';
 import type {TAddress} from '@yearn-finance/web-lib/types';
 import type {TTxResponse} from '@yearn-finance/web-lib/utils/web3/transaction';
 import type {TWriteTransaction} from './toWagmiProvider';
 
-type TApproveToken = TWriteTransaction & {
+//Because USDT do not return a boolean on approve, we need to use this ABI
+const ALTERNATE_ERC20_APPROVE_ABI = [{'constant': false, 'inputs': [{'name': '_spender', 'type': 'address'}, {'name': '_value', 'type': 'uint256'}], 'name': 'approve', 'outputs': [], 'payable': false, 'stateMutability': 'nonpayable', 'type': 'function'}] as const;
+
+/* 🔵 - Yearn Finance **********************************************************
+** isApprovedERC20 is a _VIEW_ function that checks if a token is approved for
+** a spender.
+******************************************************************************/
+export async function isApprovedERC20(
+	connector: Connector | undefined,
+	tokenAddress: TAddress,
+	spender: TAddress,
+	amount = MAX_UINT_256
+): Promise<boolean> {
+	const wagmiProvider = await toWagmiProvider(connector);
+	const result = await readContract({
+		...wagmiProvider,
+		abi: erc20ABI,
+		address: tokenAddress,
+		functionName: 'allowance',
+		args: [wagmiProvider.address, spender]
+	});
+	return (result || 0n) >= amount;
+}
+
+/* 🔵 - Yearn Finance **********************************************************
+** allowanceOf is a _VIEW_ function that returns the amount of a token that is
+** approved for a spender.
+******************************************************************************/
+type TAllowanceOf = {
+	connector: Connector | undefined,
+	tokenAddress: TAddress,
+	spenderAddress: TAddress
+}
+export async function allowanceOf(props: TAllowanceOf): Promise<bigint> {
+	const wagmiProvider = await toWagmiProvider(props.connector);
+	const result = await readContract({
+		...wagmiProvider,
+		abi: erc20ABI,
+		address: props.tokenAddress,
+		functionName: 'allowance',
+		args: [wagmiProvider.address, props.spenderAddress]
+	});
+	return result || 0n;
+}
+
+/* 🔵 - Yearn Finance **********************************************************
+** approveERC20 is a _WRITE_ function that approves a token for a spender.
+**
+** @param spenderAddress - The address of the spender.
+** @param amount - The amount of collateral to deposit.
+******************************************************************************/
+type TApproveERC20 = TWriteTransaction & {
 	spenderAddress: TAddress | undefined;
 	amount: bigint;
 };
-export async function	approveToken(props: TApproveToken): Promise<TTxResponse> {
-	assertAddress(props.contractAddress, 'contractAddress');
+export async function approveERC20(props: TApproveERC20): Promise<TTxResponse> {
 	assertAddress(props.spenderAddress, 'spenderAddress');
+	assertAddress(props.contractAddress);
+
+	props.onTrySomethingElse = async (): Promise<TTxResponse> => {
+		assertAddress(props.spenderAddress, 'spenderAddress');
+		return await handleTx(props, {
+			address: props.contractAddress,
+			abi: ALTERNATE_ERC20_APPROVE_ABI,
+			functionName: 'approve',
+			args: [props.spenderAddress, props.amount]
+		});
+	};
 
 	return await handleTx(props, {
 		address: props.contractAddress,
@@ -26,12 +89,19 @@ export async function	approveToken(props: TApproveToken): Promise<TTxResponse> {
 	});
 }
 
-type TDepositToken = TWriteTransaction & {
+/* 🔵 - Yearn Finance **********************************************************
+** deposit is a _WRITE_ function that deposits a collateral into a vault using
+** the vanilla direct deposit function.
+**
+** @app - Vaults
+** @param amount - The amount of ETH to deposit.
+******************************************************************************/
+type TDeposit = TWriteTransaction & {
 	amount: bigint;
 };
-export async function	depositToken(props: TDepositToken): Promise<TTxResponse> {
-	assertAddress(props.contractAddress, 'contractAddress');
-	assert(props.amount > 0n, 'Amount must be greater than 0');
+export async function deposit(props: TDeposit): Promise<TTxResponse> {
+	assertAddress(props.contractAddress);
+	assert(props.amount > 0n, 'Amount is 0');
 
 	return await handleTx(props, {
 		address: props.contractAddress,
@@ -41,42 +111,19 @@ export async function	depositToken(props: TDepositToken): Promise<TTxResponse> {
 	});
 }
 
-type TWithdrawToken = TWriteTransaction & {
+/* 🔵 - Yearn Finance **********************************************************
+** withdrawShares is a _WRITE_ function that withdraws a share of underlying
+** collateral from a vault.
+**
+** @app - Vaults
+** @param amount - The amount of ETH to withdraw.
+******************************************************************************/
+type TWithdrawShares = TWriteTransaction & {
 	amount: bigint;
 };
-export async function	withdrawToken(props: TWithdrawToken): Promise<TTxResponse> {
-	assertAddress(props.contractAddress, 'contractAddress');
-	assert(props.amount > 0n, 'Amount must be greater than 0');
-
-	return await handleTx(props, {
-		address: props.contractAddress,
-		abi: VAULT_ABI,
-		functionName: 'withdraw',
-		args: [props.amount]
-	});
-}
-
-type TApeInVault = TWriteTransaction & {
-	amount: bigint;
-};
-export async function	apeInVault(props: TApeInVault): Promise<TTxResponse> {
-	assertAddress(props.contractAddress, 'contractAddress');
-	assert(props.amount > 0n, 'Amount must be greater than 0');
-
-	return await handleTx(props, {
-		address: props.contractAddress,
-		abi: VAULT_ABI,
-		functionName: 'deposit',
-		args: [props.amount]
-	});
-}
-
-type TApeOutVault = TWriteTransaction & {
-	amount: bigint;
-};
-export async function	apeOutVault(props: TApeOutVault): Promise<TTxResponse> {
-	assertAddress(props.contractAddress, 'contractAddress');
-	assert(props.amount > 0n, 'Amount must be greater than 0');
+export async function withdrawShares(props: TWithdrawShares): Promise<TTxResponse> {
+	assertAddress(props.contractAddress);
+	assert(props.amount > 0n, 'Amount is 0');
 
 	return await handleTx(props, {
 		address: props.contractAddress,
